addos_v6/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==4.1.13
django-cors-headers==4.4.0
djangorestframework==3.15.1
djongo==1.3.6
dnspython==2.6.1
idna==3.10
numpy==1.24.4
packaging==24.1
pandas==2.2.3
plotly==5.24.1
pymongo==3.12.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
sqlparse==0.2.4
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
